+++ spec/dummy/client/entries/app-kitten.js
@@ -9,6 +9,7 @@ import LinkBox from 'kitten/components/box/link-box'
 
 // Buttons
 import { Button } from 'kitten/components/buttons/button'
+import { ButtonIcon } from 'kitten/components/buttons/button-icon'
 import ButtonImageWithTextAndBadge from 'kitten/components/buttons/button-image-with-text-and-badge'
 import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
 import { TagButton } from 'kitten/components/buttons/tag-button'
@@ -44,6 +45,9 @@ import GrabberIcon from 'kitten/components/icons/grabber-icon'
 import { HeartIcon } from 'kitten/components/icons/heart-icon'
 import { VisaIcon } from 'kitten/components/icons/visa-icon'
 import { LocationIcon } from 'kitten/components/icons/location-icon'
+import { FacebookIcon } from 'kitten/components/icons/facebook-icon'
+import { TwitterIcon } from 'kitten/components/icons/twitter-icon'
+import { LinkedinIcon } from 'kitten/components/icons/linkedin-icon'
 
 // Links
 import ExternalRichLink from 'kitten/components/links/external-rich-link'
@@ -130,6 +134,12 @@ import { KarlButtonIconHeartRight,
          KarlButtonIconHeartLeft
        } from 'kitten/karl/buttons/button-heart-icon-words'
 import { KarlButtonAndText } from 'kitten/karl/buttons/button'
+import { KarlButtonIconHeart,
+         KarlButtonIconMicro,
+         KarlButtonIconFacebookIcon,
+         KarlButtonIconTwitterIcon,
+         KarlButtonIconLinkedinIcon
+       } from 'kitten/karl/buttons/button-icon-heart-icon'
 import { KarlLineWithTooltip,
          KarlLineWithBigHeart } from 'kitten/karl/layout/line'
 import { KarlPaymentButton } from 'kitten/karl/buttons/payment-button'
@@ -144,6 +154,7 @@ ReactOnRails.register({
 
   // Buttons
   Button,
+  ButtonIcon,
   ButtonImageWithTextAndBadge,
   ButtonTooltipIcon,
   TagButton,
@@ -182,6 +193,9 @@ ReactOnRails.register({
   HeartIcon,
   VisaIcon,
   LocationIcon,
+  FacebookIcon,
+  TwitterIcon,
+  LinkedinIcon,
 
   // Links
   ExternalRichLink,
@@ -224,6 +238,11 @@ ReactOnRails.register({
   KarlAutoTriggerCallToActionPopover,
   KarlAutoTriggerPopover,
   KarlLoudspeakerPopover,
+  KarlButtonIconHeart,
+  KarlButtonIconMicro,
+  KarlButtonIconFacebookIcon,
+  KarlButtonIconTwitterIcon,
+  KarlButtonIconLinkedinIcon,
   KarlButtonIconHeartRight,
   KarlButtonIconHeartLeft,
   KarlButtonAndText,